Add a utility class used to manage taskbar thumbnail buttons
Add the needed ico files to use as buttons on the thumbnail, modify the
module descriptor to use the the win.taskbar module from the recently
added dependency.

Add a method named 'addButtonToStage(String, String, String,
ITaskbarButtonClickListener)' which is used to a new button to the stage
thumbnail using the specified tooltip, icon path and listener.

Add the 'addThumbnailButtons()' which adds 3 buttons to the current
stage for the pause, resume and stop recording functionalities