Implement the Recording start view pause and resume functionalities
Add resumeRecording() method and the pauseRecording() method
implementation.

Fix a bug that prevents system tray icon menuItems to set enabled and
disabled when pausing and resuming the recording. In order to do that
the MenuItems listeners got modified to use the
EventQueue.invokeLater(Runnable) method when enabling and disabling
menuItems. Add requires sentence for java.desktop to module descriptor