Fix recording duration property on paused recordings.
Modify InteractionRecorder and Recording classes to take in
consideration the time a recording was in pause when setting the
recordign duration.

To implement this, Modify the closeRecording() method from the Recording

class to receive a float parameter indicating the time the recording was
in pause. Modify code related to that method on the InteractionRecorder
class.

Modify loggin sentences on both classes.