Replace Recording date attribute by a date-time one
Modify Recording LocalDate attribute by a LocalDateTime in order
to store not only the day but the time when the Recording was created.

The LocalDateTime is an instance variable unlike the old LocalDate with
was a public class constant. Now the LocalDateTime is created using a
new public String constant indicating the DateTime format.

Modify the RecordingsRepository class accordingly to the new changes