Modify the start view to pass the recorded recording to the save view
Modify the RecordingStartViewController in order to pass the recorded
recording to the RecordingSavingView Stage after stoping the current
recording. This got implemented by using the stage.setUserData method.

Add a method to remove the app Icon from the system tray. Meant to be
used after when stoping the recording process