Implement saving fucntionality on recording saving view controller
Modify the class to receive the recorded recording after controller
initialization, also add a limit to the recording name and description
inputs.

Add a method to validate the inputs format, in case they are invalid
displays an alert suggesting to correct the given input contents.

Create the saveRecording(Event) method which is triggered after clicking
on GUI save button. This method verifies if the inputs contents are
valid by using the validateFields methods, if the given values are
invalid the method exits, else saves the recording. Then displays an
alert indicating if the recording was sucessfully saved on the database.