Implement ActionsPlayer class functionalities
Replace the RecordingPlayer class with the ActionsPlayer class.

Add a new method to 'cast' NativeInputEvents to Replayable actions, the
method name is parseReplayableAction and receives an NativeInputEvent as
parameter.

The RecordingPlayer's constructor receives an Recording instance, gets
the InputEvent and parses each one to Replayable actions and schedules
them using an ScheduledExecutorService