Fix recording NativeInputEvents execution time
Now the input event's execution time is relative to the recording start
time. To implement this changes a new instance variable called
recordingStartTime to hold the recordings start time which value is set
on the startRecording method.

Also modify addNativeEventToRecording method implementation to set the
difference between the recordings start time and the event capture time.