Modify ReplaySelectionViewController class to display stored Recordings
Now the replay selection view displays on screen the stores recording
for selection. To implement this the view fxml file and controller were
modify.

Add a new method to return the recording time stamp as a String. on
Recording class

Give acces to javafx.base to the Recording class package by modifying
the module descriptor