Modify logic to don't iterate the actions at recording replay
Instead of iterate each action of the recording to be replayed, add a
attribute to the Recording class to know if the recording has
click/movemen/keyboard/scroll events. To implement this, the
ReplayStartViewController, ActionsPlayer and ReplayableNativeRecording classes got modified to
gather this information from the recording itself intead of figuring it
at replay time.