Modify Recording, Replayable Actions and both recording and replaying process
Modify the type for the attribute inputEvents in the Recording class
from Map to Queue, this in order to increase performance, reduce app
complexity and improve code legibility. This queu stores only
ReplayableActions objects

Modify the Replayable Action class to hold a new property called
relativeExecutionTime which stores the execution time of the
ReplayableActions itselft relative to the start time of its owner
Recording. Modify ReplayableAction subclases accordingly to this change

Modify the ReplayablaNativeRecorder accordingly to change of type of the
inputEvents recordings variable and accordingly to the change in the
ReplayabaAction class types constructors

Finally also the Actions Player behavior got modified to adapt to the
previously mentioned changes