Modify ViewController behavior when showing new views
Modify the ViewController class to include a new method which
configures the Scene for the given View in to the passed Stage. This now
allows to modify the navigateToView method in order to reuse the Stage
instance from the previous View. But Without altering the loading of the
first View due to method overloading.

Improve documentation and minor style fixes on both ViewController and
StagePositioner classes