Modify Settings view to add a new config for record config saving
Modify the settings view to a new checkbox control to either enable or
disable the remember last record feature. Also improve the view design
and other minor fixes of the behavior. To implement this a new css file
was added, and both controller class and fxml file were modified.

Modify the Settings Repository class to include new method to store the
value of the setting remember last record feature